Split docker:build and docker:push completely
Build base class and splitted into DockerBuild and DockerPush
