+++ bin/command/docker_push_command.py
@@ -20,10 +20,7 @@
 
 from cleo import Output
 from jinja2 import Environment, FileSystemLoader
-from webdevops import BaseCommand
-from webdevops import Dockerfile
-from webdevops import DockerBuildTaskLoader
-from webdevops import DockerfileUtility
+from webdevops import BaseCommand, DockerPushTaskLoader
 from doit.doit_cmd import DoitMain
 
 import sys
@@ -63,7 +60,7 @@ class DockerPushCommand(BaseCommand):
         if configuration['threads'] > 1:
             doitOpts.extend(['-n', str(configuration['threads']), '-P' 'thread'])
 
-        sys.exit(DoitMain(DockerBuildTaskLoader(configuration)).run(doitOpts))
+        sys.exit(DoitMain(DockerPushTaskLoader(configuration)).run(doitOpts))
