+++ bin/webdevops/DockerBuildTaskLoader.py
@@ -24,117 +24,14 @@ import sys
 import re
 import copy
 from webdevops.BaseTaskLoader import BaseTaskLoader
+from webdevops.BaseDockerTaskLoader import BaseDockerTaskLoader
 from webdevops import DockerfileUtility
 from doit.task import dict_to_task
 
 
-class DockerBuildTaskLoader(BaseTaskLoader):
+class DockerBuildTaskLoader(BaseDockerTaskLoader):
 
-    configuration_default = {
-        'basePath': False,
-
-        'docker': {
-            'imagePrefix': '',
-            'autoLatestTag': False,
-            'fromRegExp': re.compile(ur'FROM\s+(?P<image>[^\s:]+)(:(?P<tag>.+))?', re.MULTILINE),
-            'pathRegex': False,
-            'autoPull': False,
-            'autoPullWhitelist': False,
-            'autoPullBlacklist': False,
-        },
-
-        'dockerBuild': {
-            'enabled': False,
-            'noCache': False,
-            'dryRun': False,
-            'retry': 1
-        },
-
-        'dockerPush': {
-            'enabled': False,
-            'retry': 1
-        },
-
-        'filter': {
-            'whitelist': False,
-            'blacklist': False,
-        },
-
-        'verbosity': 1,
-        'threads': 1,
-    }
-
-    dockerClient = False
-
-    def __init__(self, configuration):
-        """
-        Constrcutor
-        """
-        BaseTaskLoader.__init__(self, configuration)
-
-        # Init docker client
-        self.dockerClient = docker.from_env(assert_hostname=False)
-
-    def load_tasks(self, cmd, opt_values, pos_args):
-        """
-        DOIT task list generator
-        """
-        config = {'verbosity': self.configuration['verbosity']}
-
-        dockerfile_list = DockerfileUtility.find_dockerfiles_in_path(
-            base_path=self.configuration['basePath'],
-            path_regex=self.configuration['docker']['pathRegex'],
-            image_prefix=self.configuration['docker']['imagePrefix'],
-            whitelist=self.configuration['whitelist'],
-            blacklist=self.configuration['blacklist'],
-        )
-        dockerfile_list = self.process_dockerfile_list(dockerfile_list)
-
-        # print json.dumps(dockerfile_list, sort_keys=True, indent = 4, separators = (',', ': '));sys.exit(0);
-
-        taskList = []
-
-        # Tasks: Docker build
-        if self.configuration['dockerBuild']['enabled']:
-            taskList.extend(self.generate_tasks_docker_build(dockerfile_list))
-
-        # Tasks: Docker push
-        if self.configuration['dockerPush']['enabled']:
-            taskList.extend(self.generate_tasks_docker_push(dockerfile_list))
-
-        return taskList, config
-
-    def process_dockerfile_list(self, dockerfile_list):
-        """
-        Prepare dockerfile list with dependency and also add "auto latest tag" images
-        """
-
-        image_list = [x['image']['fullname'] for x in dockerfile_list if x['image']['fullname']]
-
-        autoLatestTagImageList = []
-
-        for dockerfile in dockerfile_list:
-            # Calculate dependency
-            dockerfile['dependency'] = False
-            if dockerfile['image']['from'] and dockerfile['image']['from'] in image_list:
-                dockerfile['dependency'] = dockerfile['image']['from']
-
-            # Process auto latest tag
-            if self.configuration['docker']['autoLatestTag'] and dockerfile['image']['tag'] == self.configuration['docker']['autoLatestTag']:
-                imageNameLatest = DockerfileUtility.generate_image_name_with_tag_latest(dockerfile['image']['fullname'])
-                if imageNameLatest not in image_list:
-                    autoLatestTagImage = copy.deepcopy(dockerfile)
-                    autoLatestTagImage['image']['fullname'] = imageNameLatest
-                    autoLatestTagImage['image']['tag'] = 'latest'
-                    autoLatestTagImage['dependency'] = dockerfile['image']['fullname']
-                    autoLatestTagImageList.append(autoLatestTagImage)
-
-        # Add auto latest tag images to dockerfile list
-        dockerfile_list.extend(autoLatestTagImageList)
-
-        return dockerfile_list
-
-    def generate_tasks_docker_build(self, dockerfileList):
+    def generate_task_list(self, dockerfileList):
         """
         Generate task list for docker build
         """
@@ -156,37 +53,6 @@ class DockerBuildTaskLoader(BaseTaskLoader):
             'name': 'FinishChain|DockerBuild',
             'title': DockerBuildTaskLoader.task_title_finish,
             'actions': [(DockerBuildTaskLoader.action_chain_finish, ['docker build'])],
-            'task_dep': [task.name for task in taskList]
-        }
-        taskList.append(dict_to_task(task))
-
-        return taskList
-
-
-    def generate_tasks_docker_push(self, dockerfile_list):
-        """
-        Generate task list for docker push
-        """
-        taskList = []
-
-
-        for dockerfile in dockerfile_list:
-            task = {
-                'name': 'DockerPush|%s' % dockerfile['image']['fullname'],
-                'title': DockerBuildTaskLoader.task_title_push,
-                'actions': [(DockerBuildTaskLoader.action_docker_push, [self.dockerClient, dockerfile, self.configuration])],
-                'task_dep': []
-            }
-
-            if self.configuration['dockerBuild']['enabled']:
-                task['task_dep'].append('FinishChain|DockerBuild')
-
-            taskList.append(dict_to_task(task))
-
-        task = {
-            'name': 'FinishChain|DockerPush',
-            'title': DockerBuildTaskLoader.task_title_finish,
-            'actions': [(DockerBuildTaskLoader.action_chain_finish, ['docker push'])],
             'task_dep': [task.name for task in taskList]
         }
         taskList.append(dict_to_task(task))
@@ -259,85 +125,8 @@ class DockerBuildTaskLoader(BaseTaskLoader):
         return build_status
 
     @staticmethod
-    def action_docker_push(docker_client, dockerfile, configuration, task):
-        """
-        Push one Docker image to registry
-        """
-        if configuration['dryRun']:
-            print '       dep: %s' % (DockerBuildTaskLoader.human_task_name_list(task.task_dep) if task.task_dep else 'none')
-            print ''
-            return
-
-        push_status = False
-        for retry_count in range(0, configuration['dockerPush']['retry']):
-            response = docker_client.push(
-                dockerfile['image']['fullname'],
-                stream=True,
-                decode=True
-            )
-            if DockerBuildTaskLoader.process_docker_client_response(response):
-                push_status = True
-                break
-            elif retry_count < (configuration['dockerBuild']['retry'] - 1):
-                print '    failed, retrying... (try %s)' % (retry_count+1)
-            else:
-                print '    failed, giving up'
-
-        return push_status
-
-    @staticmethod
-    def process_docker_client_response(response):
-        ret = True
-        last_message = False
-
-        def output_message(message, prevent_repeat=False):
-            if 'last_message' not in output_message.__dict__:
-                output_message.last_message = False
-
-            # Prevent repeating messages
-            if prevent_repeat:
-                if output_message.last_message and output_message.last_message == message:
-                    return
-                output_message.last_message = message
-            else:
-                output_message.last_message = False
-
-            sys.stdout.write(message.strip(' \t\n\r') + '\n')
-
-
-        if not response:
-            return False
-
-        for line in response:
-            # Keys
-            #   - error
-            #   - stream
-            #   - status, progressDetail, id
-            #   - progressDetail | aux [ tag, digest, size ]
-            if 'error' in line:
-                output_message(line['error'])
-                ret = False
-            if 'stream' in line:
-                output_message(line['stream'], prevent_repeat=True)
-            if 'status' in line:
-                message = line['status']
-                if 'id' in line:
-                    message += ' ' + line['id']
-                output_message(message)
-        print ''
-        return ret
-
-    @staticmethod
     def task_title_build(task):
         """
         Build task title function
         """
         return "Docker build %s" % (BaseTaskLoader.human_task_name(task.name))
-
-    @staticmethod
-    def task_title_push(task):
-        """
-        Push task title function
-        """
-        return "Docker push %s" % (BaseTaskLoader.human_task_name(task.name))
-0
